requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements/requirements.in
#
annotated-types==0.6.0
    # via pydantic
attrs==23.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.200
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v6==2.0.1
    # via aws-cdk-lib
aws-cdk-lib==2.100.0
    # via -r requirements/requirements.in
cattrs==23.1.2
    # via jsii
certifi==2023.7.22
    # via requests
charset-normalizer==3.3.0
    # via requests
constructs==10.3.0
    # via
    #   -r requirements/requirements.in
    #   aws-cdk-lib
idna==3.4
    # via requests
importlib-resources==6.1.0
    # via jsii
jsii==1.90.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-lib
    #   constructs
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-lib
    #   constructs
    #   jsii
pydantic==2.4.2
    # via -r requirements/requirements.in
pydantic-core==2.10.1
    # via pydantic
python-benedict==0.32.1
    # via -r requirements/requirements.in
python-dateutil==2.8.2
    # via jsii
python-fsutil==0.10.0
    # via python-benedict
python-slugify==8.0.1
    # via python-benedict
pyyaml==6.0.1
    # via -r requirements/requirements.in
requests==2.31.0
    # via python-benedict
six==1.16.0
    # via python-dateutil
text-unidecode==1.3
    # via python-slugify
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v6
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.8.0
    # via
    #   jsii
    #   pydantic
    #   pydantic-core
urllib3==2.0.6
    # via requests
